requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
mssql-django==1.3
pyodbc==5.0.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
django-mssql-backend==2.8.1
